Validate node support

+++ src/elements/DeployedList/DeployedList.wc.svelte
@@ -2,29 +2,32 @@
 
 <script lang="ts">
   import { onMount, onDestroy } from "svelte";
-
+  import type { IProfile } from "../../types/Profile";
+  import type { ITab } from "../../types";
   import DeployedList from "../../types/deployedList";
   import deleteContracts from "../../utils/deleteContracts";
 
   export let tab: "k8s" | "vm" | "caprover" = undefined;
 
+  // components
+  import Modal from "../../components/DeploymentModal.svelte";
+  import SelectProfile from "../../components/SelectProfile.svelte";
+  import Tabs from "../../components/Tabs.svelte";
+
   // prettier-ignore
-  const tabs = [
-    { label: "Kubernetes" },
-    { label: "Virtual Machines" },
-    { label: "Caprover" }
+  const tabs: ITab[] = [
+    { label: "Kubernetes", value: "k8s" },
+    { label: "Virtual Machines", value: "vm" },
+    { label: "Caprover", value: "caprover" }
   ];
-  let active: string = "Kubernetes";
+  let active: string = "k8s";
 
   let loading = false;
   let configed = false;
   let list: DeployedList;
-  const configs = window.configs?.baseConfig;
   const deployedStore = window.configs?.deploymentStore;
-  let profileIdx: number = 0;
 
-  $: profiles = $configs;
-  $: profile = $configs[profileIdx];
+  let profile: IProfile;
 
   function onConfigHandler() {
     configed = true;
@@ -35,8 +38,6 @@
       })
       .finally(() => (loading = false));
   }
-
-  const onSelectProfile = (e: Event) => profileIdx = (e.target as any).selectedIndex; // prettier-ignore
 
   function _reloadTab() {
     const x = active;
@@ -52,6 +53,8 @@
 
   let removed: string[] = [];
   function onRemoveHandler(key: "k8s" | "machines", name: string) {
+    const remove = window.confirm(`Are you sure u want to delete '${name}'?`);
+    if (!remove) return;
     removed = [...removed, name];
     const idx = removed.length - 1;
     deleteContracts(profile, key, name)
@@ -68,11 +71,11 @@
       });
   }
 
-  let infoToShow: string = "";
+  let infoToShow: Object;
 
   let _sub: any;
   onMount(() => {
-    _sub = deployedStore.subscribe((data) => {
+    _sub = deployedStore.subscribe(() => {
       _reloadTab();
     });
   });
@@ -82,6 +85,14 @@
   });
 </script>
 
+<SelectProfile
+  on:profile={({ detail }) => {
+    profile = detail;
+    if (detail) {
+      onConfigHandler();
+    }
+  }}
+/>
 <div style="padding: 15px;">
   <section class="box">
     <h4 class="is-size-4 mb-4">
@@ -95,71 +106,15 @@
     {#if loading}
       <p style="text-align: center; mt-2 mb-2">Loading...</p>
     {:else if !configed}
-      <form on:submit|preventDefault={onConfigHandler}>
-        <div style="display: flex; justify-content: center;">
-          <div
-            class="select mb-4"
-            style="display: flex; justify-content: flex-end;"
-          >
-            <select on:change={onSelectProfile}>
-              {#each profiles as profile, idx (idx)}
-                <option value={idx}
-                  >{#if profile.name}
-                    {profile.name}
-                  {:else}
-                    Profile {idx + 1}
-                  {/if}</option
-                >
-              {/each}
-            </select>
-          </div>
-        </div>
-        <div style="display: flex; justify-content: center;">
-          <button
-            disabled={!profile ||
-              profile.mnemonics === "" ||
-              profile.storeSecret === ""}
-            type="submit"
-            class="button is-primary"
-          >
-            List
-          </button>
-        </div>
-      </form>
-      <!--  -->
+      <p style="text-align: center; mt-2 mb-2">
+        Please activate a profile from profile manager
+      </p>
     {:else}
-      <!--  -->
-      <div style="display: flex">
-        <button
-          class="button is-primary is-outlined mr-2"
-          on:click={() => (configed = false)}
-        >
-          Back
-        </button>
+      <Tabs bind:active {tabs} />
 
-        {#if !tab}
-          <div style="width: 100%;">
-            <div class="tabs is-centered">
-              <ul>
-                {#each tabs as tab (tab.label)}
-                  <li class={active === tab.label ? "is-active" : ""}>
-                    <a
-                      href="#!"
-                      on:click|preventDefault={() => (active = tab.label)}
-                    >
-                      <span>{tab.label}</span>
-                    </a>
-                  </li>
-                {/each}
-              </ul>
-            </div>
-          </div>
-        {/if}
-      </div>
-
-      {#if (!tab && active === "Kubernetes") || tab === "k8s"}
+      {#if (!tab && active === "k8s") || tab === "k8s"}
         {#await list.loadK8s()}
-          <div class="notification is-info mt-2">&gt; Loading...</div>
+          <div class="notification is-info mt-2">Loading...</div>
         {:then rows}
           {#if rows.length}
             <div class="table-container mt-2">
@@ -186,15 +141,11 @@
                       {/if}
                       <td>{row.master.yggIP}</td>
                       <td>{row.workers}</td>
-                      <td>
+                      <td class="is-flex">
                         <button
-                          class="button is-outlined is-primary ml-2"
+                          class="button is-outlined is-primary mr-2"
                           on:click={() => {
-                            infoToShow = JSON.stringify(
-                              row.details,
-                              undefined,
-                              4
-                            );
+                            infoToShow = row.details;
                           }}
                           disabled={removed.includes(row.name)}
                         >
@@ -233,7 +184,7 @@
         {/await}
       {/if}
 
-      {#if active === "Virtual Machines" || active === "Caprover" || tab === "caprover" || tab === "vm"}
+      {#if active === "vm" || active === "caprover" || tab === "caprover" || tab === "vm"}
         {#await active === "Caprover" || tab === "caprover" ? list.loadCaprover() : list.loadVm()}
           <div class="notification is-info mt-2">&gt; Loading...</div>
         {:then rows}
@@ -252,39 +203,52 @@
                 </thead>
                 <tbody>
                   {#each rows as row, idx}
-                  {#if row.name }
-                    <tr>
-                      <th>{idx + 1}</th>
-                      <td>{row.name}</td>
-                      {#if row.publicIP}
-                        <td>{row.publicIP.ip}</td>
-                      {:else}
-                        <td>-</td>
-                      {/if}
-                      <td>{row.yggIP}</td>
-                      <td>{row.flist}</td>
-                      <td>
-                        <button
-                          class="button is-outlined is-primary  ml-2"
-                          on:click={() => {
-                            infoToShow = JSON.stringify(row, undefined, 4);
-                          }}
-                          disabled={removed.includes(row.name)}
-                        >
-                          Show Details
-                        </button>
-                        <button
-                          class={"button is-danger " +
-                            (removed.includes(row.name) ? "is-loading" : "")}
-                          on:click={() => onRemoveHandler("machines", row.name)}
-                          disabled={removed.includes(row.name)}
-                        >
-                          Delete
-                        </button>
-                      </td>
-                    </tr>
-                  {/if}
-
+                    {#if row.name}
+                      <tr>
+                        <th>{idx + 1}</th>
+                        <td>{row.name}</td>
+                        {#if row.publicIP}
+                          <td>{row.publicIP.ip}</td>
+                        {:else}
+                          <td>-</td>
+                        {/if}
+                        <td>{row.yggIP}</td>
+                        <td>{row.flist}</td>
+                        <td class="is-flex">
+                          <button
+                            class="button is-outlined is-primary"
+                            on:click={() => {
+                              infoToShow = row.details;
+                            }}
+                            disabled={removed.includes(row.name)}
+                          >
+                            Show Details
+                          </button>
+                          <button
+                            class={"button is-danger ml-2" +
+                              (removed.includes(row.name) ? "is-loading" : "")}
+                            on:click={() =>
+                              onRemoveHandler("machines", row.name)}
+                            disabled={removed.includes(row.name)}
+                          >
+                            Delete
+                          </button>
+                          {#if row.details.env && row.details.env.CAPROVER_ROOT_DOMAIN}
+                            <a
+                              class="ml-2"
+                              target="_blank"
+                              href={"http://captain." +
+                                row.details.env.CAPROVER_ROOT_DOMAIN}
+                              disabled={removed.includes(row.details.name)}
+                            >
+                              <button class="button is-link">
+                                Admin Panel
+                              </button>
+                            </a>
+                          {/if}
+                        </td>
+                      </tr>
+                    {/if}
                   {/each}
                 </tbody>
               </table>
@@ -312,19 +276,7 @@
 </div>
 
 {#if infoToShow}
-  <div class="modal is-active">
-    <div class="modal-background" />
-    <div class="modal-content">
-      <div class="box" style="white-space: pre;">
-        {infoToShow}
-      </div>
-    </div>
-    <button
-      class="modal-close is-large"
-      aria-label="close"
-      on:click={() => (infoToShow = "")}
-    />
-  </div>
+  <Modal data={infoToShow} on:closed={() => (infoToShow = null)} />
 {/if}
 
 <style lang="scss" scoped>
